requirements.txt
beautifulsoup4>=4.9.3
selenium>=4.0.0
requests>=2.25.1
feedparser>=6.0.0
python-dotenv>=0.19.0
webdriver_manager>=3.5.2
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
gunicorn>=21.2.0
